Streams syntax moved to `tofu.syntax` package.
