+++ streams/src/main/scala/tofu/syntax/streams/chunks.scala
@@ -1,4 +1,4 @@
-package tofu.streams.syntax
+package tofu.syntax.streams
 
 import tofu.streams.Chunks
