+++ streams/src/main/scala/tofu/syntax/streams/compile.scala
@@ -1,4 +1,4 @@
-package tofu.streams.syntax
+package tofu.syntax.streams
 
 import tofu.streams.{Compile, internal}
