+++ streams/src/main/scala/tofu/syntax/streams/filter.scala
@@ -1,4 +1,4 @@
-package tofu.streams.syntax
+package tofu.syntax.streams
 
 import cats.FunctorFilter
