+++ streams/src/main/scala/tofu/syntax/streams/pace.scala
@@ -1,4 +1,4 @@
-package tofu.streams.syntax
+package tofu.syntax.streams
 
 import tofu.streams.Pace
