+++ streams/src/main/scala/tofu/syntax/streams/region.scala
@@ -1,4 +1,4 @@
-package tofu.streams.syntax
+package tofu.syntax.streams
 
 import tofu.streams.Region
